Move server-only data types into `server` module
